+++ configuration.nix
@@ -84,5 +84,5 @@
     };
   };
 
-  system.stateVersion = "25.05";
+  system.stateVersion = "24.11";
 }